third_party/WebKit/LayoutTests/animations/retain-zero-percent.html

<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<div id="target"></div>
<script>
test(function() {
  var animation = target.animate([{offsetDistance: '-1%'}, {offsetDistance: '1%'}], 1);
  animation.pause();
  animation.currentTime = 0.5;
  assert_equals(getComputedStyle(target).offsetDistance, '0%');
});
</script>
